% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_numeric_genotype}
\alias{write_numeric_genotype}
\title{Write numeric dosage genotype file}
\usage{
write_numeric_genotype(snp_info, geno_mat, file, sep = ",")
}
\arguments{
\item{snp_info}{A data.frame or data.table of SNP metadata with columns
\code{SNP}, \code{CHR}, \code{POS}, \code{REF}, \code{ALT}.}

\item{geno_mat}{Numeric matrix of dosage values (0/1/2/NA), SNPs x samples.}

\item{file}{Output file path.}

\item{sep}{Field separator. Default \code{","}.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write numeric dosage genotype file
}
